requirements.txt
streamlit
pymysql
python-dotenv
openai
tzdata
google-generativeai